backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
httpx==0.28.1
python-dotenv==1.0.1
drf-spectacular==0.27.2
psycopg[binary]==3.2.12
